Change CampaignCode text to CartDiscountLabel
Fixes an issue with new gift cards; when a campaign code and a gift card
are applied at the same time, the CartDiscount tag will return all
discounts and gift cards as a single total value in the compability
mode. Therefore the total's label should be printed with the tag instead
of dictionary text because of the tag can return more specific indicator
that the total contains both campaign code and gift card as a single
amount.